:recycle: refactor : move JoinedParticipantsDto.java to service layer
